requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.3.6
requests==2.32.3
langchain==0.3.4
openai==1.75.0
python-dotenv==1.1.0
google-auth==2.39.0
google-api-core==2.24.2
google-cloud-speech==2.32.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
proto-plus==1.26.1
protobuf==5.29.4
rsa==4.9.1
pinecone-client==6.0.0
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
click==8.1.8
MarkupSafe==3.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.4.0
langgraph==0.2.20
langchain-openai==0.2.1
beautifulsoup4==4.12.3
tweepy==4.14.0
